Refactored dev-component styling
Moved all component into own folders and moved thier styling from scss/ directory into each components folder.
Also added own scss file for container-card and moved %accents to _button.scss
