+++ client/src/dev-components/container-card/container-card.scss
@@ -1,0 +1,45 @@
+@import '../../scss/compulsory';
+
+@import '../../scss/atoms-organisms/button';
+@import '../../scss/atoms-organisms/tooltip';
+
+.instance-card {
+  flex: 1;
+  padding: $spacer;
+  margin: $spacer;
+  border-radius: $border-radius-default;
+  box-shadow: 0 0 10px $palette-shadow-default;
+
+  &:not(:first-of-type) {
+    margin-top: 0;
+  }
+
+  .header-container {
+    display: flex;
+    justify-content: space-between;
+    flex-flow: row-reverse wrap;
+
+    h5 {
+      text-transform: capitalize;
+      margin-right: auto;
+    }
+
+    .header-button-container {
+      display: flex;
+      justify-content: space-between;
+
+      button {
+        @extend %icon-button;
+      }
+    }
+  }
+
+
+  a {
+    @extend %anchor-button;
+
+    &.accent {
+      @extend %accent;
+    }
+  }
+}
