+++ client/src/dev-components/create-runner/create-runner.scss
@@ -2,8 +2,8 @@
 @import '../../scss/colors';
 @import '../../scss/mixins';
 
-@import '../atoms-organisms/tooltip';
-@import './create.scss';
+@import '../../scss/atoms-organisms/tooltip';
+@import '../create/create';
 
 .runner-view-container {
   display: flex;
