+++ client/src/dev-components/create/create.scss
@@ -3,7 +3,7 @@
 @import '../../scss/atoms-organisms/button';
 @import '../../scss/atoms-organisms/forms';
 
-@import './dashboard-view.scss';
+@import '../dashboard-view/dashboard-view';
 @import '../../feedbacker-components/reactions/reactions';
 
 .center-center-block {
