+++ client/src/dev-components/dashboard-view/dashboard-view.scss
@@ -1,0 +1,47 @@
+@import '../../scss/compulsory';
+
+@import '../../scss/atoms-organisms/button';
+@import '../../scss/atoms-organisms/tooltip';
+
+.top-section {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+
+  padding: $spacer * 3;
+  box-shadow: 0 0 10px $palette-shadow-light;
+}
+
+.create-button {
+  @extend %anchor-button-accent;
+  align-self: flex-end;
+}
+
+.instances-container {
+  display: flex;
+  flex-flow: column;
+  justify-content: center;
+  align-items: center;
+  padding: $spacer * 2;
+
+  h2 {
+    margin-top: $spacer * 3;
+    margin-bottom: $spacer * 2;
+  }
+
+  .reverse {
+    display: flex;
+    flex-direction: column-reverse;
+  }
+}
+
+.button-container {
+  display: flex;
+  justify-content: space-between;
+  margin-top: $spacer;
+
+  > *:not(:first-of-type) {
+    margin-left: $spacer;
+  }
+}
+
