+++ client/src/feedbacker-components/survey-edit/survey-edit-card.scss
@@ -4,7 +4,7 @@
 
 @import '../../scss/atoms-organisms/button';
 
-@import '../../scss/views/dashboard-view';
+@import '../../dev-components/dashboard-view/dashboard-view';
 @import '../../scss/atoms-organisms/forms';
 
 .survey-edit-card {
