+++ client/src/scss/atoms-organisms/_button.scss
@@ -1,5 +1,16 @@
 @import '../../scss/compulsory';
 
+%accent {
+  box-shadow: 0 2px 5px $palette-shadow-default;
+  background: $palette-accent;
+  color: $palette-lighter;
+}
+
+%anchor-button-accent {
+  @extend %anchor-button;
+  @extend %accent;
+}
+
 @mixin reset-button() {
   border: none;
   margin: 0;
