+++ client/src/site.js
@@ -12,9 +12,9 @@ import {
 } from 'react-router-dom'
 
 import { setupPersist } from './persist'
-import Dashboard from './dev-components/dashboard-view'
-import Create from './dev-components/create'
-import Build from './dev-components/build-view'
+import Dashboard from './dev-components/dashboard-view/dashboard-view'
+import Create from './dev-components/create/create'
+import Build from './dev-components/build-view/build-view'
 import InvalidContainer from './invalid-container'
 import { prepareReactRoot } from './shadowDomHelper'
 import { setUsers, subscribeUpdateUsers, setUserName, showCookieToast } from './globals'
@@ -23,7 +23,7 @@ import { setPersistData } from './actions'
 // Styles
 import styles from './scss/_base.scss'
 import './scss/atoms-organisms/_toast.scss'
-import CreateRunner from './dev-components/create-runner'
+import CreateRunner from './dev-components/create-runner/create-runner'
 
 const css = classNames.bind(styles)
